Update README.md
You will be creating a button menu that has a hover state and also allows you to select a button when you click on it. The final project can be seen in this Codepen.

Step 1
Setup a project on your local computer and then open it in your code editor. In the root folder create files for index.html styles.css and index.js

Step 2
Copy and paste the code into their corresponding files